refactor: BottomSheetReservationTicketArg로 클래스명 변경
