+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/ticketreserve/bottomsheet/BottomSheetReservationTicketArg.kt
@@ -1,10 +1,10 @@
-package com.festago.festago.presentation.ui.ticketreserve
+package com.festago.festago.presentation.ui.ticketreserve.bottomsheet
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class ReservationTicketArg(
+data class BottomSheetReservationTicketArg(
     val id: Int,
     val remainAmount: Int,
     val ticketType: String,
