+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/ticketreserve/bottomsheet/TicketReserveBottomItem.kt
@@ -1,8 +1,6 @@
 package com.festago.festago.presentation.ui.ticketreserve.bottomsheet
 
-import com.festago.festago.presentation.ui.ticketreserve.ReservationTicketArg
-
 data class TicketReserveBottomItem(
-    val ticket: ReservationTicketArg,
+    val ticket: BottomSheetReservationTicketArg,
     val isSelected: Boolean = false,
 )
