+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/ticketreserve/bottomsheet/TicketReserveBottomSheetFragment.kt
@@ -6,7 +6,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import com.festago.festago.databinding.FragmentTicketReserveBottomSheetBinding
-import com.festago.festago.presentation.ui.ticketreserve.ReservationTicketArg
 import com.festago.festago.presentation.ui.ticketreserve.TicketReserveViewModel
 import com.festago.festago.presentation.util.getParcelableArrayListCompat
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
@@ -44,7 +43,7 @@ class TicketReserveBottomSheetFragment : BottomSheetDialogFragment() {
             getString(KEY_STAGE_START_TIME)?.let { startTime ->
                 binding.stageStartTime = startTime
             }
-            getParcelableArrayListCompat<ReservationTicketArg>(KEY_ITEMS)?.let {
+            getParcelableArrayListCompat<BottomSheetReservationTicketArg>(KEY_ITEMS)?.let {
                 ticketTypeAdapter.submitList(it.map(::TicketReserveBottomItem))
             }
         }
@@ -69,7 +68,7 @@ class TicketReserveBottomSheetFragment : BottomSheetDialogFragment() {
 
         fun newInstance(
             stageStartTime: String,
-            items: List<ReservationTicketArg>,
+            items: List<BottomSheetReservationTicketArg>,
         ) = TicketReserveBottomSheetFragment().apply {
             arguments = Bundle().apply {
                 putString(KEY_STAGE_START_TIME, stageStartTime)
